bug fixes and modular v0.9
