+++ backend/requirements.txt
@@ -1,9 +1,45 @@
+asttokens==2.2.1
+backcall==0.2.0
+blinker==1.6.2
+certifi==2023.7.22
+charset-normalizer==3.2.0
+click==8.1.7
+colorama==0.4.6
+comm==0.1.4
+decorator==5.1.1
+et-xmlfile==1.1.0
+executing==1.2.0
 Flask==2.3.3
-flask-cors==4.0.0
+Flask-Cors==4.0.0
+idna==3.4
 ipython==8.14.0
 ipywidgets==8.1.0
-pandas==2.0.3
-plotly==5.16.1
-Requests==2.31.0
-schedule==1.2.0
+itsdangerous==2.1.2
+jedi==0.19.0
+Jinja2==3.1.2
+jupyterlab-widgets==3.0.8
+MarkupSafe==2.1.3
+matplotlib-inline==0.1.6
+numpy==1.25.2
 openpyxl==3.1.2
+packaging==23.1
+pandas==2.0.3
+parso==0.8.3
+pickleshare==0.7.5
+plotly==5.16.1
+prompt-toolkit==3.0.39
+pure-eval==0.2.2
+Pygments==2.16.1
+python-dateutil==2.8.2
+pytz==2023.3
+requests==2.31.0
+schedule==1.2.0
+six==1.16.0
+stack-data==0.6.2
+tenacity==8.2.3
+traitlets==5.9.0
+tzdata==2023.3
+urllib3==2.0.4
+wcwidth==0.2.6
+Werkzeug==2.3.7
+widgetsnbextension==4.0.8
